requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
langchain==0.3.25
langchain-openai==0.3.16
python-dotenv==1.1.0
pydantic==2.11.4
pydantic-settings==2.9.1
faiss-cpu==1.11.0
jira==3.8.0
langchain-community==0.3.23